Down Strike Animation Improved Further & Raycast Fun
Blades: Used a butter sprite for the first frame in the down strike animation. Down see blade handle for a moment. Also, allowed for better smoothing. Also, I improved the animation positioning itself, so it starts right side of center, goes down center screen, ends left of center.

Flail: Used a better sprite for the first frame in the flail downstrike animation. This allowed for a much smoother overall animation for the down strike on the flail. Flail looks like the ball actually flies out from in front of your.

Messing with raycasting more for better hit lineup with animations.